Move shared navigation related components into separate file
