+++ frontend/src/citizen-frontend/navigation/AttentionIndicator.tsx
