+++ frontend/src/citizen-frontend/navigation/CityLogo.tsx
